Working implementation
This is working locally and on Azure although it has outdated references to all static JSON sites and the groupJSON site, so those need to be updated.

Also all HTML / JavaScript was tossed into server.js in order to get this to work. A future user story might be to separate these again.